(NB: Database schema/view/functions all changed)
Updated database so that version specific information is now stored in
the wb_versions table. It keeps the timestamp of the update, the
username of the user, the appid of the app and the client IP of the
user/app.

The has reduced some redundancy, as the timestamp and user were
previously stored in the wb_triple_vers table with each triple. That
table now just has the version, triple_order and triple.

Refreshing the database is required for all databases for this to work,
otherwise the functions etc will not match up, and your server will
error.